fixed dialogs opened from the column menu
