+++ frontend/.eslintrc.json
@@ -12,7 +12,8 @@
     "react/prop-types": 0,
     "no-console": 0,
     "jsx-a11y/label-has-for": 0,
-    "jsx-a11y/no-autofocus": 0
+    "jsx-a11y/no-autofocus": 0,
+    "react/display-name": 0
   },
   "plugins": ["react", "import", "jsx-a11y", "cypress"],
   "parser": "babel-eslint",
