+++ frontend/src/components/dialogs/DeleteColumnDialog.js
@@ -1,5 +1,4 @@
-import React, { useState, useContext } from "react";
-import DeleteIcon from "@material-ui/icons/Delete";
+import React, { useContext } from "react";
 import {
   Dialog,
   DialogActions,
@@ -7,28 +6,15 @@ import {
   DialogContentText,
   DialogTitle,
   Button,
-  withMobileDialog,
-  MenuItem,
-  ListItemIcon,
-  ListItemText
+  withMobileDialog
 } from "@material-ui/core";
 
 import { DELETE_COLUMN } from "../../constants/eventNames";
 import { BoardContext } from "../../context/BoardContext";
-import { DELETE_COLUMN_BUTTON } from "../../constants/testIds";
 
 function DeleteColumnDialog(props) {
-  const { columnId, fullScreen } = props;
-  const [open, setOpen] = useState(false);
+  const { isOpen, closeDialog, columnId, fullScreen } = props;
   const { boardId, socket } = useContext(BoardContext);
-
-  function openDialog() {
-    setOpen(true);
-  }
-
-  function closeDialog() {
-    setOpen(false);
-  }
 
   function handleClick() {
     socket.emit(DELETE_COLUMN, columnId, boardId);
@@ -36,39 +22,31 @@ function DeleteColumnDialog(props) {
   }
 
   return (
-    <>
-      <MenuItem button onClick={openDialog} data-testid={DELETE_COLUMN_BUTTON}>
-        <ListItemIcon>
-          <DeleteIcon fontSize="small" />
-        </ListItemIcon>
-        <ListItemText inset primary="Delete Column" />
-      </MenuItem>
-      <Dialog
-        fullScreen={fullScreen}
-        open={open}
-        onClose={closeDialog}
-        aria-labelledby="alert-delete-column-dialog"
-        aria-describedby="alert-delete-column-dialog-description"
-      >
-        <DialogTitle id="alert-delete-column-dialog">
-          {"Delete this column?"}
-        </DialogTitle>
-        <DialogContent>
-          <DialogContentText id="alert-delete-column-dialog-description">
-            You are about to delete this column. If you are sure, then click on
-            the delete button.
-          </DialogContentText>
-        </DialogContent>
-        <DialogActions>
-          <Button onClick={closeDialog} color="primary">
-            Cancel
-          </Button>
-          <Button onClick={handleClick} color="primary" autoFocus>
-            Delete
-          </Button>
-        </DialogActions>
-      </Dialog>
-    </>
+    <Dialog
+      fullScreen={fullScreen}
+      open={isOpen}
+      onClose={closeDialog}
+      aria-labelledby="alert-delete-column-dialog"
+      aria-describedby="alert-delete-column-dialog-description"
+    >
+      <DialogTitle id="alert-delete-column-dialog">
+        {"Delete this column?"}
+      </DialogTitle>
+      <DialogContent>
+        <DialogContentText id="alert-delete-column-dialog-description">
+          You are about to delete this column. If you are sure, then click on
+          the delete button.
+        </DialogContentText>
+      </DialogContent>
+      <DialogActions>
+        <Button onClick={closeDialog} color="primary">
+          Cancel
+        </Button>
+        <Button onClick={handleClick} color="primary" autoFocus>
+          Delete
+        </Button>
+      </DialogActions>
+    </Dialog>
   );
 }
